Remove redundant file referencing ListAdapter
